Acoustic indices included in web app csv export

+++ snaw-backend/api.py
@@ -12,7 +12,7 @@ from analysis_driver import run_driver
 import traceback
 import random
 import shutil
-from keras.models import load_model
+
 
 UPLOAD_FOLDER = 'instance/upload/'
 ALLOWED_EXTENSIONS = {'wav'}
@@ -23,7 +23,6 @@ app.config['UPLOAD_FOLDER'] = UPLOAD_FOLDER
 app.secret_key = 'secret key'
 app.config['SESSION_TYPE'] = 'filesystem'
 DEBUG_FLAG = True
-
 '''
 ###------------------------------------------------------###
 App Routing: '/'
@@ -171,8 +170,11 @@ def run_analysis():
 
 @app.route('/removeUserFolder', methods=['GET', 'POST'])
 def closeUserFolder():
-    shutil.rmtree('instance/upload/user' + session['id'])
+    if os.path.isdir('instance/upload/user'+session['id']):
+        shutil.rmtree('instance/upload/user' + session['id'])
     return redirect('', 204)
+
+
 
 #print('Starting Flask!')
 app.run(debug=True)
